backend: Move Speed::mask() implementation to Cynthion backend.
