+++ src/backend/mod.rs
@@ -30,16 +30,6 @@ impl Speed {
             Low => "Low (1.5Mbps)",
         }
     }
-
-    pub fn mask(&self) -> u8 {
-        use Speed::*;
-        match self {
-            Auto => 0b0001,
-            Low  => 0b0010,
-            Full => 0b0100,
-            High => 0b1000,
-        }
-    }
 }
 
 pub struct InterfaceSelection {
